test: 토스 결제 승인 API 테스트

+++ src/main/java/com/moabam/api/dto/payment/ConfirmTossPaymentResponse.java
@@ -14,7 +14,7 @@ public record ConfirmTossPaymentResponse(
 	String orderId,
 	String orderName,
 	PaymentStatus status,
-	Long totalAmount,
+	int totalAmount,
 	LocalDateTime approvedAt
 ) {
